* spearated MapTemplates and template management from contexts * sources fixed
