+++ web/client/epics/context.js
@@ -7,31 +7,19 @@
  */
 
 import { Observable } from 'rxjs';
-import { get, pick, isObject, isArray, find, cloneDeep } from 'lodash';
 
 import { LOCATION_CHANGE } from 'connected-react-router';
 
-import MapUtils from '../utils/MapUtils';
 import { getResource, getResourceIdByName, getResourceDataByName } from '../api/persistence';
-import Api from '../api/GeoStoreDAO';
-import { error as showError } from '../actions/notifications';
 import { pluginsSelectorCreator } from '../selectors/localConfig';
 import { isLoggedIn } from '../selectors/security';
-import { mapSelector } from '../selectors/map';
-import { layersSelector, groupsSelector } from '../selectors/layers';
-import { backgroundListSelector } from '../selectors/backgroundselector';
-import { textSearchConfigSelector } from '../selectors/searchconfig';
-import { mapOptionsToSaveSelector } from '../selectors/mapsave';
-import { templatesSelector, currentContextSelector, mapTemplatesLoadedSelector } from '../selectors/context';
 
-import { LOAD_CONTEXT, LOAD_FINISHED, loadContext, loading, setContext, setResource, contextLoadError, loadFinished, setMapTemplatesLoaded,
-    setTemplateData, setTemplateLoading, SET_CURRENT_CONTEXT, CONTEXT_LOAD_ERROR, OPEN_MAP_TEMPLATES_PANEL, MERGE_TEMPLATE,
-    REPLACE_TEMPLATE } from '../actions/context';
-import { loadMapConfig, configureMap, MAP_CONFIG_LOADED, MAP_CONFIG_LOAD_ERROR } from '../actions/config';
+import { LOAD_CONTEXT, LOAD_FINISHED, loadContext, loading, setContext, setResource, contextLoadError, loadFinished,
+    SET_CURRENT_CONTEXT, CONTEXT_LOAD_ERROR } from '../actions/context';
+import { clearMapTemplates } from '../actions/maptemplates';
+import { loadMapConfig, MAP_CONFIG_LOADED, MAP_CONFIG_LOAD_ERROR } from '../actions/config';
 import { changeMapType } from '../actions/maptype';
-import { setControlProperty } from '../actions/controls';
 import { LOGIN_SUCCESS, LOGOUT } from '../actions/security';
-
 
 import { wrapStartStop } from '../observables/epics';
 import ConfigUtils from '../utils/ConfigUtils';
@@ -98,6 +86,7 @@ const errorToMessageId = (name, e, getState = () => {}) => {
 export const loadContextAndMap = (action$, { getState = () => { } } = {}) =>
     action$.ofType(LOAD_CONTEXT).switchMap(({ mapId, contextName }) =>
         Observable.merge(
+            Observable.of(clearMapTemplates()),
             getResourceIdByName('CONTEXT', contextName)
                 .switchMap(id => createContextFlow(id, action$, getState)).catch(e => {throw new ContextError(e); }),
             (mapId ? Observable.of(null) : getResourceDataByName('CONTEXT', contextName))
@@ -145,114 +134,3 @@ export const handleLoginLogoutContextReload = action$ => {
         action$.ofType(LOAD_CONTEXT)
     ).switchMap(([, args]) => Observable.of(loadContext(args)));
 };
-
-export const openMapTemplatesPanelEpic = (action$, store) => action$
-    .ofType(OPEN_MAP_TEMPLATES_PANEL)
-    .switchMap(() => {
-        const state = store.getState();
-        const currentContext = currentContextSelector(state);
-        const templates = templatesSelector(state);
-        const mapTemplatesLoaded = mapTemplatesLoadedSelector(state);
-
-        const makeFilter = () => ({
-            OR: {
-                FIELD: templates.map(template => ({
-                    field: ['ID'],
-                    operator: ['EQUAL_TO'],
-                    value: [template.id]
-                }))
-            }
-        });
-
-        const extractThumbnail = (resource) => {
-            const attribute = get(resource, 'Attributes.attribute', {});
-            const attributes = isArray(attribute) ? attribute : [attribute];
-            return get(find(attributes, ({name}) => name === 'thumbnail'), 'value');
-        };
-
-        return Observable.of(setControlProperty('mapTemplates', 'enabled', true, true)).concat(!mapTemplatesLoaded ?
-            Observable.defer(() => Api.searchListByAttributes(makeFilter(), {
-                params: {
-                    includeAttributes: true
-                }
-            }, '/resources/search/list'))
-                .switchMap((data) => {
-                    const resourceObj = get(data, 'ResourceList.Resource', []);
-                    const resources = isArray(resourceObj) ? resourceObj : [resourceObj];
-                    const newTemplates = resources.map(resource => ({
-                        ...pick(resource, 'id', 'name', 'description'),
-                        thumbnail: extractThumbnail(resource),
-                        dataLoaded: false,
-                        loading: false
-                    }));
-                    return Observable.of(setContext({...currentContext, templates: newTemplates}), setMapTemplatesLoaded(true));
-                })
-                .catch(err => Observable.of(setMapTemplatesLoaded(true, err))) :
-            Observable.empty()
-        );
-    });
-
-export const mergeTemplateEpic = (action$, store) => action$
-    .ofType(MERGE_TEMPLATE)
-    .switchMap(({id}) => {
-        const state = store.getState();
-        const templates = templatesSelector(state);
-        const template = find(templates, t => t.id === id);
-
-        return (template.dataLoaded ? Observable.of(template.data) :
-            Observable.defer(() => Api.getData(id))).switchMap(data => {
-            if (isObject(data) && data.map !== undefined) {
-                const map = mapSelector(state);
-                const layers = layersSelector(state);
-                const groups = groupsSelector(state);
-                const backgrounds = backgroundListSelector(state);
-                const textSearchConfig = textSearchConfigSelector(state);
-                const additionalOptions = mapOptionsToSaveSelector(state);
-
-                const currentConfig = MapUtils.saveMapConfiguration(map, layers, groups, backgrounds, textSearchConfig, additionalOptions);
-                const newConfig = MapUtils.mergeMapConfigs(currentConfig, data);
-
-                return Observable.of(...(!template.dataLoaded ? [setTemplateData(id, data)] : []), configureMap(cloneDeep(newConfig), null));
-            }
-
-            return !template.dataLoaded ? Observable.of(setTemplateData(id, data)) : Observable.empty();
-        }).let(wrapStartStop(
-            setTemplateLoading(id, true),
-            setTemplateLoading(id, false),
-            e => {
-                const messageId = errorToMessageId('template', e.originalError, store.getState);
-                return Observable.of(showError({
-                    title: 'context.errors.template.title',
-                    message: messageId,
-                    position: "tc",
-                    autoDismiss: 5
-                }));
-            }
-        ));
-    });
-
-export const replaceTemplateEpic = (action$, store) => action$
-    .ofType(REPLACE_TEMPLATE)
-    .switchMap(({id}) => {
-        const state = store.getState();
-        const templates = templatesSelector(state);
-        const template = find(templates, t => t.id === id);
-
-        return (template.dataLoaded ? Observable.of(template.data) :
-            Observable.defer(() => Api.getData(id))).switchMap(data => Observable.of(
-            ...(!template.dataLoaded ? [setTemplateData(id, data)] : []),
-            ...(isObject(data) && data.map !== undefined ? [configureMap(cloneDeep(data), null)] : [])
-        )).let(wrapStartStop(
-            setTemplateLoading(id, true),
-            setTemplateLoading(id, false),
-            e => {
-                const messageId = errorToMessageId('template', e.originalError, store.getState);
-                return Observable.of(showError({
-                    title: 'context.errors.template.title',
-                    message: messageId,
-                    position: "tc",
-                    autoDismiss: 5
-                }));
-            }
-        ));
-    });
